Changed the name of the new filter from IP Classification to IP User.
