+++ repose-aggregator/components/filters/filter-bundle/src/main/application/WEB-INF/web-fragment.xml
@@ -137,7 +137,7 @@
         <filter-class>org.openrepose.filters.urlextractortoheader.UrlExtractorToHeaderFilter</filter-class>
     </filter>
     <filter>
-        <filter-name>ip-classification</filter-name>
-        <filter-class>org.openrepose.filters.ipclassification.IpClassificationFilter</filter-class>
+        <filter-name>ip-user</filter-name>
+        <filter-class>org.openrepose.filters.ipuser.IpUserFilter</filter-class>
     </filter>
 </web-fragment>
