--- repose-aggregator/components/filters/ip-classification/src/main/resources/META-INF/schema/examples/ip-classification.cfg.xml
@@ -1,59 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  _=_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=
-  Repose
-  _-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-
-  Copyright (C) 2010 - 2015 Rackspace US, Inc.
-  _-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
-  -->
-
-<ip-classification xmlns="http://docs.openrepose.org/repose/ip-classification/v1.0">
-    <!--
-     Specify what header to put the user value in. This is the default.
-     It will always add an additional header, not clobbering the existing value.
-    -->
-    <user-header name="X-PP-User" quality="0.4"/>
-
-    <!--
-     Specify what header to put the group name in. This is the default.
-     It will always add an additional header, not clobbering the existing value.
-    -->
-    <group-header name="X-PP-Groups" quality="0.4"/>
-
-    <!--
-    Group match criterion follow the CIDR rules: https://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing
-    You can specify CIDR expressions in either IPv4, IPv6, or mixed.
-    The groups should be priority ordered as the first match wins.
-    If any group has a match, then the name is applied to the specified header and matching stops.
-    If nothing matches, then nothing happens.
-    A default catch all is not supplied.
-    If you want to guarantee at least one group will match, then you must define a catch all group.
-    The names of the groups must be unique.
-    -->
-    <group name="sample-group">
-        <cidr-ip>192.168.1.0/24</cidr-ip>
-        <cidr-ip>192.168.0.1/32</cidr-ip>
-    </group>
-    <group name="sample-ipv6-group">
-        <cidr-ip>2001:db8::/48</cidr-ip>
-    </group>
-    <group name="bolth-group">
-        <cidr-ip>10.10.220.0/24</cidr-ip>
-        <cidr-ip>2001:1938:80:bc::1/64</cidr-ip>
-    </group>
-    <group name="ipv4-match-all">
-        <cidr-ip>0.0.0.0/0</cidr-ip>
-    </group>
-</ip-classification>
