--- repose-aggregator/functional-tests/spock-functional-test/src/test/configs/features/filters/ipclassification/otherconfig/ip-classification.cfg.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  _=_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=
-  Repose
-  _-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-
-  Copyright (C) 2010 - 2015 Rackspace US, Inc.
-  _-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-  =_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_=_
-  -->
-
-<ip-classification xmlns="http://docs.openrepose.org/repose/ip-classification/v1.0">
-    <user-header quality="0.7"/>
-    <group-header quality="0.5"/>
-    <group name="other-group">
-        <cidr-ip>10.6.51.192/33</cidr-ip>
-    </group>
-    <group name="local-group">
-        <cidr-ip>127.0.0.1/32</cidr-ip>
-    </group>
-</ip-classification>
